+++ dsr_bringup/CHANGELOG.rst
@@ -2,6 +2,10 @@
 Changelog for package dsr_bringup
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
+1.1.0 (20-01-2025)
+------------------
+* Added ros_action_name as parameter.
+
 1.0.0 (16-01-2025)
 ------------------
 * Initial release.